backend/pyproject.toml
[project]
name = "neighborly-backend"
version = "0.1.0"
description = "Neighborly backend API"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.133.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.41.0",
    "playwright>=1.50.0",
    "supabase>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
